Archive legacy resources and add third-party notices
Move legacy resource files into DOCS/HISTORICAL/legacy-resources and add a README and notes for migration reference. Add SnakeYAML license and consolidated THIRD-PARTY notices, and update pom.xml to package LICENSE and third-party license files under META-INF. Fix an integer overflow by switching combined memory calculation to long and add a test for it; remove an unused import. Improve ProcessSupervisor output thread keep-alive, enable core thread timeout, expose outputWorkerCount, and add tests for thread retirement and handoff. Update README and todo to reflect these changes.

+++ src/main/java/net/slimelabs/slslite/config/ConfigurationValidator.java
@@ -90,7 +90,7 @@ public final class ConfigurationValidator {
                     "Managed lobby blueprint not found: "
                             + config.lobby().registry() + "/" + config.lobby().server()
             ));
-            int requiredMemory = limboMemory + lobbyBlueprint.memoryLimitMiB();
+            long requiredMemory = (long) limboMemory + lobbyBlueprint.memoryLimitMiB();
             if (requiredMemory > config.totalMemoryMiB()) {
                 throw new ConfigurationException(
                         "Managed lobby and SLS-Limbo require " + requiredMemory
